imdb-clone
all major datafetch endpoints are added

+++ IMDB-CLONE-USERSERVICE/src/main/java/org/roybond007/model/helper/RatingReference.java
@@ -18,7 +18,7 @@ public class RatingReference {
 
 	private String id;
 	
-	private double rating;
+	private long rating;
 	
 	private long timestamp;
